requirements.txt
python-telegram-bot
transformers
torch
torchvision
torchaudio
pymongo
sounddevice
numpy
librosa
requests
pyaudio
pydub
ffmpeg-downloader
SpeechRecognition
googletrans-py
easydict
matplotlib
prettytable
protobuf
scikit-image